Added files via upload
Hey guys, this is the basic html homepage (landing page) that is already on the server.  It can be viewed right now at picotekdesign.com/science_companion.html.  I know Sarah mentioned wanting to work on the landing page and that's totally cool.  This was already a task that was on Trello before she mentioned that.  So I'm totally open for a redesign or whatever.  This will just be what's up on the server for now and I wanted to upload this file in case any of you want to add to or modify it.  Remember, this file can be saved in any directory on your computer to be edited and viewed and we can later push any files to the server later.  I'll be working on getting the other subject pages added to Git here in the next day or so.  They will be very similar to the template I posted this past weekend.